chrome/content/mochitest/tests/test_policy.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Content policy tests</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>        

  <script type="application/x-javascript;version=1.7" src="common.js"></script>
  <script type="application/x-javascript;version=1.7" src="../httpd.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="visibility: hidden;">
    <iframe id="frame" onload="doRunTest()"></iframe>
  </div>

  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    let abp = Components.classes["@mozilla.org/adblockplus;1"].createInstance().wrappedJSObject;

    let tests = [
      [
        '<img src="test.gif">',
        "http://localhost:1234/test.gif", "image", "localhost", false
      ],
      [
        '<img src="http://127.0.0.1:1234/test.gif">',
        "http://127.0.0.1:1234/test.gif", "image", "localhost", true
      ],
      [
        '<input type="image" src="test.gif">',
        "http://localhost:1234/test.gif", "image", "localhost", false
      ],
      [
        '<div style="background-image: url(test.gif)"></div>',
        "http://localhost:1234/test.gif", "background", "localhost", false
      ],
      [
        '<div style="cursor: url(test.gif), pointer"></div>',
        "http://localhost:1234/test.gif", "background", "localhost", false
      ],
      [
        <ol>
          <li style="list-style-image: url(test.gif)">foo</li>
        </ol>,
        "http://localhost:1234/test.gif", "background", "localhost", false
      ],
      [
        '<style>div:before { content: url(test.gif); }</style><div>foo</div>',
        "http://localhost:1234/test.gif", "background", "localhost", false
      ],
      [
        '<embed type="image/gif" src="test.gif"></embed>',
        "http://localhost:1234/test.gif", "object", "localhost", false
      ],
      [
        '<object type="image/gif" data="test.gif"></object>',
        "http://localhost:1234/test.gif", "object", "localhost", false
      ],
      [
        <window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
          <image src="test.gif"/>
        </window>,
        "http://localhost:1234/test.gif", "image", "localhost", false
      ],
      [
        <window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
          <image style="list-style-image: url(test.gif)"/>
        </window>,
        "http://localhost:1234/test.gif", "background", "localhost", false
      ],
      [
        <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
          <image xlink:href="test.gif"/>
        </svg>,
        "http://localhost:1234/test.gif", "image", "localhost", false
      ],
      [
        <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
          <filter>
            <feImage xlink:href="test.gif"/>
          </filter>
        </svg>,
        "http://localhost:1234/test.gif", "image", "localhost", false
      ],
      [
        '<' + 'script src="test.js"><' + '/script>',
        "http://localhost:1234/test.js", "script", "localhost", false
      ],
      [
        <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
          <script xlink:href="test.js"/>
        </svg>,
        "http://localhost:1234/test.js", "script", "localhost", false
      ],
      [
        '<link rel="stylesheet" type="text/css" href="test.css">',
        "http://localhost:1234/test.css", "stylesheet", "localhost", false
      ],
      [
        '<?xml-stylesheet href="test.css" type="text/css"?><window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"/>',
        "http://localhost:1234/test.css", "stylesheet", "localhost", false
      ],
    ];

    if (window.navigator.mimeTypes["application/x-shockwave-flash"] && window.navigator.mimeTypes["application/x-shockwave-flash"].enabledPlugin)
    {
      tests.push([
        '<embed type="application/x-shockwave-flash" src="test.swf"></embed>',
        "http://localhost:1234/test.swf", "object", "localhost", false
      ],
      [
        '<object type="application/x-shockwave-flash" data="test.swf"></object>',
        "http://localhost:1234/test.swf", "object", "localhost", false
      ]);
    }

    if (window.navigator.mimeTypes["application/x-java-applet"] && window.navigator.mimeTypes["application/x-java-applet"].enabledPlugin)
    {
      // Note: this could use some improvement but Gecko will fail badly with more complicated tests (bug 364400)
      // Note: <applet> is not on the list because it shows some weird async behavior (data is loaded after page load in some strange way)
      tests.push([
        '<embed type="application/x-java-applet" code="test.class" src="test.class"></embed>',
        "http://localhost:1234/test.class", "object", "localhost", false
      ],
      [
        '<object type="application/x-java-applet" data="test.class"></object>',
        "http://localhost:1234/test.class", "object", "localhost", false
      ]);
    }

    if (compareGeckoVersion("1.9.1b3") >= 0)
    {
      tests.push([
        '<style type="text/css">@font-face { font-family: Test; src: url("test.otf"); } html { font-family: Test; }</style>',
        "http://localhost:1234/test.otf", "font", "localhost", false
      ]);
    }

    // TODO: '<' + 'script>new Worker("test.js");<' + '/script>'
    // Web workers need special treatment, no way to ensure that web worker loads before the page is loaded.

    let currentTest = 0;
    let currentStage = 0;

    let policyHits = [];
    let serverHit = false;
    let stageDescriptions = {
      1: "running without filters",
      2: "running with filter %S",
      3: "running with filter %S and site exception",
      4: "running with filter %S and exception not applicable to sites",
      5: "running with filter %S and exception only applicable for frame",
    };

    function onPolicyHit(wnd, type, data, location)
    {
      if (wnd != top || type != "add" || location.location == "http://localhost:1234/test")
        return;
      policyHits.push([wnd, type, data, location]);
    }

    function runNextTest()
    {
      currentStage++;
      if (currentStage > 5)
      {
        currentTest++;
        currentStage = 1;
      }
      if (currentTest >= tests.length)
      {
        SimpleTest.finish();
        return;
      }

      let [body, expectedURL, expectedType, expectedDomain, expectedThirdParty] = tests[currentTest];

      abp.blacklistMatcher.clear();
      abp.whitelistMatcher.clear();

      if (currentStage > 1)
        abp.blacklistMatcher.add(abp.Filter.fromText(expectedURL));
      if (currentStage == 3)
        abp.whitelistMatcher.add(abp.Filter.fromText("@@|chrome://mochikit/"));
      if (currentStage == 4)
        abp.whitelistMatcher.add(abp.Filter.fromText("@@|chrome://mochikit/$~document"));
      if (currentStage == 5)
        abp.whitelistMatcher.add(abp.Filter.fromText("@@|http://localhost:1234/test|"));

      serverHit = false;
      server.registerPathHandler(expectedURL.replace(/http:\/\/[^\/]+/, ""), function(metadata, response)
      {
        serverHit = true;
        response.setStatusLine("1.1", "404", "Not Found");
      });

      policyHits = [];
      document.getElementById("frame").contentWindow.location.href = "http://localhost:1234/test";
    }

    function doRunTest()
    {
      if (currentStage == 0)
        return;

      let [body, expectedURL, expectedType, expectedDomain, expectedThirdParty] = tests[currentTest];

      let expectedStatus = "allowed";
      if (currentStage == 3)
        is(policyHits.length, 0, "Number of policy hits for " + body + " with site whitelisting");
      else
      {
        let stageDescription = stageDescriptions[currentStage];
        if (stageDescription.indexOf("%S") >= 0)
          stageDescription = stageDescription.replace("%S", expectedURL);

        is(policyHits.length, 1, "Number of policy hits for " + body + " (" + stageDescription + ")");
        if (policyHits.length == 1)
        {
          let [wnd, type, data, hit] = policyHits[0];

          is(hit.location, expectedURL, "Checking request URL");

          expectedStatus = (currentStage == 1 ? "allowed" : "blocked");
          let actualStatus = (hit.filter ? "blocked" : "allowed");

          is(actualStatus, expectedStatus, "Checking whether request was blocked (" + stageDescription + ")");
          is (hit.typeDescr.toLowerCase(), expectedType, "Checking whether request type was determined correctly (" + stageDescription + ")");
          is (hit.thirdParty, expectedThirdParty, "Checking whether third-party flag was set correctly (" + stageDescription + ")");
          is (hit.docDomain, expectedDomain, "Checking whether document domain was set correctly (" + stageDescription + ")");
        }
      }

      server.registerPathHandler(expectedURL.replace(/http:\/\/[^\/]+/, ""), null);
      is(serverHit, expectedStatus == "allowed", "Checking whether request was received by server");

      runNextTest();
    }

    function start()
    {
      abp.DataContainer.addListener(onPolicyHit);

      window.addEventListener("unload", stop, false);
      server.start(1234);

      server.registerPathHandler("/test", function(metadata, response)
      {
        let body = tests[currentTest][0];
        if (body instanceof XML)
          body = body.toXMLString();
        response.setStatusLine("1.1", "200", "OK");

        let contentType = "text/html";
        if (body.indexOf("there.is.only.xul") >= 0)
          contentType = "application/vnd.mozilla.xul+xml";
        else if (body.indexOf("2000/svg") >= 0)
          contentType = "image/svg+xml";
        response.setHeader("Content-Type", contentType);

        response.bodyOutputStream.write(body, body.length);
      });

      abp.prefs.enabled = true;
      runNextTest();
    }

    function stop()
    {
      abp.DataContainer.removeListener(onPolicyHit);

      server.stop();

      abp.filterStorage.triggerSubscriptionObservers("reload", abp.filterStorage.subscriptions);
      abp.prefs.enabled = oldEnabled;
    }

    let server = new nsHttpServer();
    let oldEnabled = abp.prefs.enabled;
    SimpleTest.waitForExplicitFinish();
    addLoadEvent(start);
  </script>
  </pre>
</body>
</html>
